requirements.txt
streamlitg
sklearn
nltk
bs4
distance
re